+++ decidim-forms/app/forms/decidim/forms/answer_form.rb
@@ -49,7 +49,22 @@ module Decidim
         choices.select(&:body)
       end
 
+      def conditions_fulfilled?
+        question.display_conditions.all? do |condition|
+          answer = question.questionnaire.answers.find_by(question: condition.condition_question)
+          condition.fulfilled?(answer)
+        end
+      end
+
       private
+
+      def mandatory_body?
+        question.mandatory_body? if conditions_fulfilled?
+      end
+
+      def mandatory_choices?
+        question.mandatory_choices? if conditions_fulfilled?
+      end
 
       def grouped_choices
         selected_choices.group_by(&:matrix_row_id).values