+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Due to [\#5553](https://github.com/decidim/decidim/pull/5553), SSL is turned on 
 
 ### Added
 
+- **decidim-forms**: Manage conditions to hide or show questions in surveys [\#5798](https://github.com/decidim/decidim/pull/5798)
 - **decidim-forms**: New question type "Matrix" [\#5876](https://github.com/decidim/decidim/pull/5876)
 - **decidim-proposals**: Add images to proposal cards [\#5640](https://github.com/decidim/decidim/pull/5640)
 - **decidim-api**: Added documentation to use the API (newcomers friendly). [\#5582](https://github.com/decidim/decidim/pull/5582)